requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.9.1
asteval==1.0.6
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.39.5
botocore==1.39.5
bw-migrations==0.2
bw2calc==2.1
bw2data==4.5
bw2io==0.9.11
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
constructive_geometries==1.0
country_converter==1.3.1
cryptography==45.0.5
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.2.18
diff-match-patch==20241021
Django==5.2.4
django-import-export==4.3.8
django-redis==6.0.0
django-storages==1.14.6
django-unfold==0.62.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2025.7.0
greenlet==3.2.3
idna==3.10
ijson==3.4.0
import-export==0.3.1
iniconfig==2.1.0
isodate==0.7.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
linear-tsv==1.1.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
matrix_utils==0.6
morefs==0.2.2
mrio-common-metadata==0.2.1
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
Pint==0.24.4
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyecospold==4.0.0
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
PyPrind==2.11.3
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
randonneur==0.6.2
randonneur_data==0.6
RapidFuzz==3.13.0
rdflib==7.1.4
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rfc3986==2.0.0
rpds-py==0.26.0
s3transfer==0.13.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stats_arrays==0.7
stripe==12.3.0
structlog==25.4.0
tableschema==1.21.0
tablib==3.8.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unfold==1.2.3
unicodecsv==0.14.1
urllib3==2.5.0
voluptuous==0.15.2
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
wurst==0.4
xlrd==2.0.2
xlsxwriter==3.2.5
